+++ wasm-crypto-init/asm.ts
@@ -3,7 +3,9 @@ import type { InitFn } from 'https://deno.land/x/polkadot/wasm-bridge/types.ts';
 import type { WasmCryptoInstance } from './types.ts';
 
 import { createWasmFn } from 'https://deno.land/x/polkadot/wasm-bridge/mod.ts';
-import { asmJsInit } from 'https://deno.land/x/polkadot/wasm-crypto-asmjs/mod.ts';
+import * as mod from 'https://deno.land/x/polkadot/wasm-crypto-asmjs/mod.ts';
+
+const { asmJsInit } = mod;
 
 export { packageInfo } from './packageInfo.ts';