+++ wasm-util/packageInfo.ts
@@ -1,3 +1,3 @@
 
 
-export const packageInfo = { name: '@polkadot/wasm-util', path: new URL(import.meta.url).pathname, type: 'deno', version: '7.4.1' };
+export const packageInfo = { name: '@polkadot/wasm-util', path: new URL(import.meta.url).pathname, type: 'deno', version: '7.5.1' };